requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
attrs==25.3.0
binance==0.3.45
ccxt==4.4.86
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
dateparser==1.2.1
dotenv==0.9.9
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pycryptodome==3.23.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
websocket-client>=1.6.0
python-telegram-bot>=20.0
yarl==1.20.0
pytest
orjson >= 3.10,<4

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
pyqtgraph>=0.13.0
PyQt5>=5.15.0
